+++ test.mjs
@@ -1475,22 +1475,30 @@ SELECT
 -- test null-case handling-behavior
 DROP TABLE IF EXISTS __tmp1;
 CREATE TEMP TABLE __tmp1 (val REAL);
-SELECT 1 AS id, quantile(val, 0.5) AS val FROM __tmp1;
+SELECT
+        1 AS id,
+        median(val) AS mdn,
+        quantile(val, 0.5) AS qnt,
+        stdev(val) AS std
+    FROM __tmp1;
                     `)
                 })
             )[0];
-            assertJsonEqual(valActual, [{id: 1, val: null}]);
+            assertJsonEqual(
+                valActual,
+                [{id: 1, mdn: null, qnt: null, std: null}]
+            );
         }());
         await Promise.all([
             [
-                [[], -99, null],
-                [[], 0, null],
-                [[], 0.5, null],
-                [[], 1, null],
-                [[], 99, null],
-                [[null, null, 1, 1, 2, 3, 4], 0.5, 2],
-                [[null, null, 1, 2, 3, 4], 0.5, 2.5],
-                [[null], 0.5, null]
+                [[], -99, undefined],
+                [[], 0, undefined],
+                [[], 0.5, undefined],
+                [[], 1, undefined],
+                [[], 99, undefined],
+                [[undefined, undefined, 1, 1, 2, 3, 4], 0.5, 2],
+                [[undefined, undefined, 1, 2, 3, 4], 0.5, 2.5],
+                [[undefined], 0.5, undefined]
             ],
             [
                 [[], -99, 1],
@@ -1539,29 +1547,72 @@ SELECT 1 AS id, quantile(val, 0.5) AS val FROM __tmp1;
         ].flat().map(async function ([
             data, kk, valExpected
         ]) {
+            let avg = 0;
+            let data2;
+            let valExpectedMdn;
+            let valExpectedStd = 0;
+            data2 = data.filter(function (elem) {
+                return elem !== undefined;
+            }).sort();
+            valExpectedMdn = (
+                data2.length % 2 === 0
+                ? 0.5 * (
+                    data2[0.5 * data2.length - 1] + data2[0.5 * data2.length]
+                )
+                : data2[0.5 * (data2.length - 1)]
+            );
+            data2.forEach(function (elem) {
+                avg += elem;
+            });
+            avg *= 1 / data2.length;
+            data2.forEach(function (elem) {
+                valExpectedStd += (elem - avg) ** 2;
+            });
+            valExpectedStd = Math.round(1_000_000 * (
+                data2.length <= 0
+                ? undefined
+                // : data2.length === 1
+                // ? 0
+                : Math.sqrt(valExpectedStd / (data2.length - 1))
+            ));
             await Promise.all([
                 data,
                 Array.from(data).reverse()
             ].map(async function (data) {
-                let valActual = noop(
+                let valActual;
+                valActual = noop(
                     await dbExecAsync({
                         bindList: {
                             tmp1: JSON.stringify(data)
                         },
                         db,
                         sql: (`
-SELECT quantile(value, ${kk}) AS val FROM JSON_EACH($tmp1);
+SELECT
+        median(value) AS mdn,
+        quantile(value, ${kk}) AS qnt,
+        ROUND(1000000 * stdev(value)) AS std
+    FROM JSON_EACH($tmp1);
 -- test null-case handling-behavior
-SELECT quantile(value, ${kk}) AS val FROM JSON_EACH($tmp1) WHERE 0;
+SELECT quantile(value, ${kk}) AS qnt FROM JSON_EACH($tmp1) WHERE 0;
                         `)
                     })
-                )[0][0].val;
-                assertJsonEqual(valActual, valExpected, {
-                    data,
-                    kk,
+                )[0][0];
+                assertJsonEqual(
                     valActual,
-                    valExpected
-                });
+                    {
+                        mdn: valExpectedMdn,
+                        qnt: valExpected ?? null,
+                        std: valExpectedStd
+                    },
+                    {
+                        data,
+                        kk,
+                        valActual,
+                        valExpected,
+                        valExpectedMdn,
+                        valExpectedStd
+                    }
+                );
             }));
         }));
     });